backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
requests
sqlalchemy
passlib
python-jose[cryptography]
python-multipart